delete schemas for db megre to models

+++ internal/domain/models/session.go
@@ -15,7 +15,7 @@ type Session struct {
 
 type SessionsCollection struct {
 	Data  []Session `json:"data"`
-	Page  uint      `json:"page"`
-	Size  uint      `json:"size"`
-	Total uint      `json:"total"`
+	Page  uint64    `json:"page"`
+	Size  uint64    `json:"size"`
+	Total uint64    `json:"total"`
 }
